web/requirements_bak.txt
Package                   Version
------------------------- --------------
aiohappyeyeballs          2.4.6
aiohttp                   3.11.12
aiosignal                 1.3.2
alembic                   1.14.1
annotated-types           0.7.0
anyio                     4.8.0
APScheduler               3.11.0
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 3.0.0
async-lru                 2.0.4
asyncer                   0.0.8
attrs                     25.1.0
babel                     2.17.0
backoff                   2.2.1
beautifulsoup4            4.13.3
bleach                    6.2.0
cachetools                5.5.1
Camelot                   12.6.29
certifi                   2025.1.31
cffi                      1.17.1
chardet                   5.2.0
charset-normalizer        3.4.1
click                     8.1.8
cloudpickle               3.1.1
colorlog                  6.9.0
comm                      0.2.2
contourpy                 1.3.1
cryptography              43.0.3
cycler                    0.12.1
dataclasses-json          0.6.7
datasets                  2.21.0
debugpy                   1.8.12
decorator                 5.1.1
defusedxml                0.7.1
dill                      0.3.8
diskcache                 5.6.3
distro                    1.9.0
dnspython                 2.7.0
dspy                      2.6.4
dspy-ai                   2.6.6
Elixir                    0.7.1
email_validator           2.2.0
et_xmlfile                2.0.0
executing                 2.2.0
faiss-cpu                 1.10.0
fastapi                   0.115.8
fastapi-sso               0.16.0
fastjsonschema            2.21.1
filelock                  3.17.0
fonttools                 4.56.0
fqdn                      1.5.1
frozenlist                1.5.0
fsspec                    2024.6.1
greenlet                  3.1.1
gunicorn                  22.0.0
h11                       0.14.0
htmlmin                   0.1.12
httpcore                  1.0.7
httpx                     0.28.1
httpx-sse                 0.4.0
huggingface-hub           0.28.1
idna                      3.10
ijson                     3.3.0
ImageHash                 4.3.2
importlib_metadata        8.6.1
ipykernel                 6.29.5
ipython                   8.32.0
ipywidgets                8.1.5
isoduration               20.11.0
jedi                      0.19.2
Jinja2                    3.1.5
jiter                     0.8.2
joblib                    1.4.2
json_repair               0.39.0
json5                     0.10.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
jupyter_client            8.6.3
jupyter_core              5.7.2
jupyter-events            0.12.0
jupyter-lsp               2.2.5
jupyter_server            2.15.0
jupyter_server_terminals  0.5.3
jupyterlab                4.3.5
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
jupyterlab_widgets        3.0.13
kiwisolver                1.4.8
langchain                 0.3.19
langchain-community       0.3.17
langchain-core            0.3.35
langchain-huggingface     0.1.2
langchain-text-splitters  0.3.6
langfuse                  2.59.3
langsmith                 0.3.8
litellm                   1.61.8
magicattr                 0.1.6
Mako                      1.3.9
MarkupSafe                2.1.5
marshmallow               3.26.1
matplotlib                3.10.0
matplotlib-inline         0.1.7
missingno                 0.5.2
mistune                   3.1.2
mpmath                    1.3.0
multidict                 6.1.0
multimethod               2.0
multiprocess              0.70.16
mypy-extensions           1.0.0
nbclient                  0.10.2
nbconvert                 7.16.6
nbformat                  5.10.4
nest-asyncio              1.6.0
networkx                  3.4.2
notebook_shim             0.2.4
numpy                     1.26.4
nvidia-cublas-cu12        12.1.3.1
nvidia-cuda-cupti-cu12    12.1.105
nvidia-cuda-nvrtc-cu12    12.1.105
nvidia-cuda-runtime-cu12  12.1.105
nvidia-cudnn-cu12         9.1.0.70
nvidia-cufft-cu12         11.0.2.54
nvidia-curand-cu12        10.3.2.106
nvidia-cusolver-cu12      11.4.5.107
nvidia-cusparse-cu12      12.1.0.106
nvidia-nccl-cu12          2.20.5
nvidia-nvjitlink-cu12     12.8.61
nvidia-nvtx-cu12          12.1.105
oauthlib                  3.2.2
openai                    1.64.0
openpyxl                  3.1.5
optuna                    4.2.1
orjson                    3.10.15
overrides                 7.7.0
packaging                 24.2
pandas                    2.2.3
pandas-profiling          3.2.0
pandocfilters             1.5.1
parso                     0.8.4
pbr                       6.1.1
pexpect                   4.9.0
phik                      0.12.4
pillow                    11.1.0
pip                       25.0
platformdirs              4.3.6
prometheus_client         0.21.1
prompt_toolkit            3.0.50
propcache                 0.2.1
psutil                    7.0.0
ptyprocess                0.7.0
pure_eval                 0.2.3
pyarrow                   19.0.1
pycparser                 2.22
pydantic                  2.10.6
pydantic_core             2.27.2
pydantic-settings         2.7.1
Pygments                  2.19.1
PyJWT                     2.10.1
PyNaCl                    1.5.0
pyparsing                 3.2.1
pytesseract               0.3.13
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
python-json-logger        3.2.1
python-multipart          0.0.18
pytz                      2025.1
PyWavelets                1.8.0
PyYAML                    6.0.2
pyzmq                     26.2.1
redis                     5.2.1
referencing               0.36.2
regex                     2024.11.6
requests                  2.32.3
requests-toolbelt         1.0.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.22.3
rq                        2.1.0
safetensors               0.5.2
scikit-learn              1.6.1
scipy                     1.15.2
seaborn                   0.13.2
Send2Trash                1.8.3
sentence-transformers     3.4.1
setuptools                75.8.0
six                       1.17.0
sniffio                   1.3.1
soupsieve                 2.6
SQLAlchemy                2.0.38
sqlalchemy-migrate        0.11.0
sqlparse                  0.5.3
stack-data                0.6.3
starlette                 0.45.3
sympy                     1.13.3
tangled-up-in-unicode     0.2.0
Tempita                   0.6.0
tenacity                  9.0.0
terminado                 0.18.1
threadpoolctl             3.5.0
tiktoken                  0.9.0
tinycss2                  1.4.0
tokenizers                0.21.0
torch                     2.4.0
torchtext                 0.18.0
tornado                   6.4.2
tqdm                      4.67.1
traitlets                 5.14.3
transformers              4.49.0
triton                    3.0.0
types-python-dateutil     2.9.0.20241206
typing_extensions         4.12.2
typing-inspect            0.9.0
tzdata                    2025.1
tzlocal                   5.3
ujson                     5.10.0
uri-template              1.3.0
urllib3                   2.3.0
uvicorn                   0.29.0
uvloop                    0.21.0
visions                   0.7.4
wcwidth                   0.2.13
webcolors                 24.11.1
webencodings              0.5.1
websocket-client          1.8.0
wheel                     0.45.1
widgetsnbextension        4.0.13
wrapt                     1.17.2
xlrd                      0.7.1
xlwt                      0.7.2
xxhash                    3.5.0
yarl                      1.18.3
zipp                      3.21.0
zstandard                 0.23.0
